+++ clang/lib/AST/CMakeLists.txt
@@ -55,6 +55,7 @@ add_clang_library(clangAST
   ExprConstant.cpp
   ExprCXX.cpp
   ExprObjC.cpp
+  ExprUtils.cpp
   ExternalASTMerger.cpp
   ExternalASTSource.cpp
   FormatString.cpp
